% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{ck_collections}
\alias{ck_collections}
\title{This function checks to make sure that the collections listed are correct}
\usage{
ck_collections(collections)
}
\arguments{
\item{collections}{List of collections to check}
}
\description{
This function checks to make sure that the collections listed are correct
}
\keyword{Note:}
\keyword{are}
\keyword{be}
\keyword{but}
\keyword{collections}
\keyword{configuration}
\keyword{file}
\keyword{hard-coded}
\keyword{internal}
\keyword{moved}
\keyword{now}
\keyword{right}
\keyword{to}
\keyword{will}
