% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{ck_date}
\alias{ck_date}
\title{This function checks that dates are correctly formatted
And it converts dates to a standard format}
\usage{
ck_date(dates)
}
\arguments{
\item{dates}{List of dates to check}
}
\description{
This function checks that dates are correctly formatted
And it converts dates to a standard format
}
