% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{ck_entities}
\alias{ck_entities}
\title{This function will makes sure a valid entity is given}
\usage{
ck_entities(url, entity.type)
}
\arguments{
\item{url}{URL to pass back with entity.type info added}

\item{entity.type}{Object to check}
}
\description{
This function will makes sure a valid entity is given
}
