% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{ck_fields}
\alias{ck_fields}
\title{Check and validate fields before sending to API endpoint}
\usage{
ck_fields(fields)
}
\arguments{
\item{fields}{List of fields to check}
}
\value{
Error message if invalid fields are found, otherwise NULL
}
\description{
Check and validate fields before sending to API endpoint
}
