% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{ck_list}
\alias{ck_list}
\title{This function will ensure a valid format for options that allow multiple values such as collections, ids, fields}
\usage{
ck_list(x)
}
\arguments{
\item{x}{Object to check}
}
\description{
This function will ensure a valid format for options that allow multiple values such as collections, ids, fields
}
