% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{ck_topics}
\alias{ck_topics}
\title{This function will makes sure a valid entity is given}
\usage{
ck_topics(topics.values)
}
\arguments{
\item{topics.values}{Topic IDs to validate}
}
\description{
This function will makes sure a valid entity is given
}
\keyword{internal}
