% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{ck_wikidata}
\alias{ck_wikidata}
\title{This function will makes sure a valid entity is given}
\usage{
ck_wikidata(entity.values)
}
\arguments{
\item{entity.values}{Wikidata ID to validate}
}
\description{
This function will makes sure a valid entity is given
}
