% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{config.ent}
\alias{config.ent}
\title{This function configures the entity values to make it compatible with the API}
\usage{
config.ent(entity.value)
}
\arguments{
\item{entity.value}{The text or list of entities to search}
}
\description{
This function configures the entity values to make it compatible with the API
}
