% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{config.topics}
\alias{config.topics}
\title{This function configures the topics values to make it compatible with the API}
\usage{
config.topics(topics.value)
}
\arguments{
\item{topics.value}{The text or list of topics to search}
}
\description{
This function configures the topics values to make it compatible with the API
}
\keyword{internal}
