% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{configsearch}
\alias{configsearch}
\title{This function configures the search text to make it compatible with the API}
\usage{
configsearch(s.text, or)
}
\arguments{
\item{s.text}{The text or list of texts to search}

\item{or}{Whether to join text with 'or' or with 'and'}
}
\description{
This function configures the search text to make it compatible with the API
}
\keyword{internal}
