% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{find.entity.id}
\alias{find.entity.id}
\title{This function finds the API listing for a specific entity}
\usage{
find.entity.id(entity.type, value = NULL)
}
\arguments{
\item{entity.type}{The type of entity to search}

\item{value}{The name or names of the entity to search}
}
\description{
This function finds the API listing for a specific entity
}
