% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{find.topics.id}
\alias{find.topics.id}
\title{This function finds the API listing for a topic containing a term}
\usage{
find.topics.id(value = NULL)
}
\arguments{
\item{value}{The name or names of the entity to search}
}
\description{
This function finds the API listing for a topic containing a term
}
