% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{fmt_collnames}
\alias{fmt_collnames}
\title{This function formats the collection names option when it is specified}
\usage{
fmt_collnames(url, coll.name)
}
\arguments{
\item{url}{API URL}

\item{coll.name}{Name of collection or list of names of collections}
}
\description{
This function formats the collection names option when it is specified
}
