% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{fmt_fields}
\alias{fmt_fields}
\title{Format fields for API request}
\usage{
fmt_fields(url, fields)
}
\arguments{
\item{url}{Base API URL}

\item{fields}{List of fields to include}
}
\description{
Format fields for API request
}
\keyword{internal}
