% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_date.R
\name{hlapi_date}
\alias{hlapi_date}
\title{History Lab API - Date search}
\usage{
hlapi_date(date=NULL,start.date=NULL,end.date=NULL,fields=NULL,coll.name=NULL,
limit = 25,run=TRUE,...)
}
\arguments{
\item{date}{Specific date to search}

\item{start.date}{Date range for the search}

\item{end.date}{Date range for the search}

\item{fields}{Specific History Lab fields to return}

\item{coll.name}{Name of History Lab collection to search}

\item{limit}{Number of results to return (Default = 25)}

\item{run}{Run query (Default) or return API URL only}

\item{...}{Not used}
}
\value{
Returns a json object (if run = TRUE) or API URL (if run=FALSE)
}
\description{
This program allows a date or date range search using the History Lab API.
Options to limit the results include collection and fields.
}
\examples{
hlapi_date(date='1980-09-15', fields=c("doc_id","classification","title"), coll.name="frus")
hlapi_date(start.date='1947-01-01', end.date='12/01/1948', limit=100)
hlapi_date(start.date='1947-01-01', end.date='12/01/1948', fields=c("doc_id","classification","title", "entities"), limit=10)
}
