% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlapi_search.R
\name{hlapi_search}
\alias{hlapi_search}
\title{History Lab API - Full text search}
\usage{
hlapi_search(s.text, fields=NULL, or = FALSE, start.date=NULL,end.date=NULL,
coll.name=NULL, limit = 25,run=TRUE,...)
}
\arguments{
\item{s.text}{Text to search--can be a single word, multiword phrase, or multiple variants}

\item{fields}{Specific History Lab fields to return}

\item{or}{Join text list with 'or' rather than 'and'}

\item{start.date}{Date range for the search}

\item{end.date}{Date range for the search}

\item{coll.name}{History Lab collection to search}

\item{limit}{Number of results to return (Default = 25)}

\item{run}{Run query (Default) or return API URL only}

\item{...}{Not used}
}
\value{
Returns a json object (if run = TRUE) or API URL (if run=FALSE)
}
\description{
This program allows a full text search using the History Lab API.
In addition to a search term, collection and start.date and end.date are required fields.
}
\examples{
hlapi_search('asean', coll.name=c('cfpf','frus'),
start.date="1974-01-01", end.date="1979-12-31")
\dontrun{
hlapi_search(c('udeac','asean'), run=TRUE)
}
hlapi_search(c('udeac','asean'), or=TRUE)
hlapi_search(c('league of nations'),  limit = 10, fields=c('doc_id','title'))
}
