% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlapi_topics.R
\name{hlapi_topics}
\alias{hlapi_topics}
\title{History Lab API - Topics search}
\usage{
hlapi_topics( topics.value, fields=NULL, coll.name=NULL,
date=NULL,start.date=NULL, end.date=NULL, or=TRUE, run=TRUE, limit = 25, ...)
}
\arguments{
\item{topics.value}{The topic ID value to be searched}

\item{fields}{Specific History Lab fields to return}

\item{coll.name}{History Lab collection to search}

\item{date}{Specific date to search}

\item{start.date}{Date range for the search}

\item{end.date}{Date range for the search}

\item{or}{Join topics list with 'or' (Default) rather than 'and'}

\item{run}{Run query (Default) or return API URL only}

\item{limit}{Number of results to return (Default = 25)}

\item{...}{Not used}
}
\value{
Returns a json object (if run = TRUE) or API URL (if run=FALSE)
}
\description{
This program allows a search of the History Lab API by topic.
Because topics are unique to collections, the function should be used with specific collections.
}
\examples{
find.topics.id('nuclear') # nuclear topic ID is 78 in frus
hlapi_topics(topics.value='78', coll.name='frus')
find.topics.id('import') # topic ID 56 in WB
find.topics.id('invest') # topic ID 51 in WB
hlapi_topics(topics.value=c("51","56"), coll.name='worldbank', fields=c('doc_id','title','authored','topic_names'), or=FALSE)

}
