% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{hlresults}
\alias{hlresults}
\title{This function sends a request to the HL API}
\usage{
hlresults(url)
}
\arguments{
\item{url}{API URL address to call}
}
\description{
This function sends a request to the HL API
}
