% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{ck_collections}
\alias{ck_collections}
\title{This function checks to make sure that the collections listed are correct}
\usage{
ck_collections(collections)
}
\arguments{
\item{collections}{List of collections to check
Note: right now collections are hard-coded but will be moved to configuration file}
}
\description{
This function checks to make sure that the collections listed are correct
}
