% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histlabapi_utils.R
\name{ck_fields}
\alias{ck_fields}
\title{This function checks to make sure that the fields listed are correct
And it gets the field list ready to send to API endpoint}
\usage{
ck_fields(fields)
}
\arguments{
\item{fields}{List of fields to check
Note: right now fields are hard-coded but will be moved to configuration file}
}
\description{
This function checks to make sure that the fields listed are correct
And it gets the field list ready to send to API endpoint
}
