% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlapi_entity.R
\name{hlapi_entity}
\alias{hlapi_entity}
\title{History Lab API - Entity search}
\usage{
hlapi_entity(entity.type=c('topics','persons','countries'), country.value, topic.value,
person.value, fields=NULL, coll.name=NULL,date=NULL,start.date=NULL, end.date=NULL,
summary= FALSE, orlist, run=NULL, limit = 25, ...)
}
\arguments{
\item{entity.type}{The type of entity to search (countries, topics, and/or persons)}

\item{country.value}{The specific country entity to be searched}

\item{topic.value}{The specific topic entity to be searched}

\item{person.value}{The specific person entity to be searched}

\item{fields}{Specific History Lab fields to return}

\item{coll.name}{History Lab collection to search}

\item{date}{Specific date to search}

\item{start.date}{Date range for the search}

\item{end.date}{Date range for the search}

\item{summary}{Aggregate the results by entity.type (Default = FALSE)}

\item{orlist}{Join entity list with 'or' rather than 'and'}

\item{run}{Run query or return API URL only (Default)}

\item{limit}{Number of results to return (Default = 25)}

\item{...}{Not used}
}
\value{
Returns a json object (if run = TRUE) or API URL (if run=FALSE)
}
\description{
This program allows a search of the History Lab API by type of entity (countries, topics, or persons).
}
