% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlapi_overview.R
\name{hlapi_overview}
\alias{hlapi_overview}
\title{History Lab API - Overview}
\usage{
hlapi_overview(aspect=NULL,  sort = NULL,  coll.name=NULL,
entity.type=NULL, limit=250, run=TRUE,...)
}
\arguments{
\item{aspect}{Type of overview}

\item{sort}{Sort in (A)scending or (D)escending order. Default = No sort}

\item{coll.name}{History Lab collection to search}

\item{entity.type}{Entity type to return (PERSON, ORG, LOC, GOVT, and OTHER)}

\item{limit}{Number of results to return (Default = 250)}

\item{run}{Run query (Default) or return API URL only}

\item{...}{Not used}
}
\value{
Returns a json object (if run = TRUE) or API URL (if run=FALSE)
}
\description{
This function allows an overview query of the History Lab API.
It will provide a list of fields and collections available to search as well as an overview of the entities and topics  in a collection.


Options to limit the results include collection and limit.
}
\examples{
hlapi_overview(aspect='collections')
hlapi_overview(aspect='entities', entity.type = 'PERSON', sort = "D")
hlapi_overview(aspect='topics',coll.name='frus')
hlapi_overview(aspect='topics',coll.name='frus', run=FALSE)

}
